opt: move catalog.Table utilities to utils.go
This makes it easier to focus on just the interfaces.

Release note: None
